# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "bluebell 社区论坛后端接口文档",
        "title": "bluebell 社区论坛 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/community": {
            "get": {
                "description": "查询所有社区（id、name），以列表形式返回",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "社区模块"
                ],
                "summary": "获取所有社区",
                "responses": {
                    "200": {
                        "description": "服务繁忙",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/community/{id}": {
            "get": {
                "description": "根据社区ID获取社区详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "社区模块"
                ],
                "summary": "获取社区详情",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "社区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或服务繁忙",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户登录接口，返回 user_id、user_name 和 JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或用户名密码错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/post": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建帖子接口，需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子模块"
                ],
                "summary": "创建帖子",
                "parameters": [
                    {
                        "description": "帖子参数（title、content、community_id）",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Post"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或未登录",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "description": "根据帖子ID获取帖子详情（含作者、社区信息、投票数）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子模块"
                ],
                "summary": "获取帖子详情",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或服务繁忙",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "分页获取帖子列表，按创建时间倒序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子模块"
                ],
                "summary": "帖子列表（按时间倒序）",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务繁忙",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/posts2": {
            "get": {
                "description": "按时间或分数排序获取帖子列表，支持社区ID筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子模块"
                ],
                "summary": "帖子列表（支持排序和社区筛选）",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "time",
                            "score"
                        ],
                        "type": "string",
                        "default": "time",
                        "description": "排序方式: time 或 score",
                        "name": "order",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "社区ID（0或不传表示全局）",
                        "name": "community_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或服务繁忙",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "用户注册接口，传入用户名、密码、确认密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUpParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或用户已存在",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/vote": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "对帖子进行投票（赞成=1，反对=-1，取消=0），需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "投票模块"
                ],
                "summary": "帖子投票",
                "parameters": [
                    {
                        "description": "投票参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamVoteData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误、未登录、重复投票或超出投票时间",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ResCode": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParam",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeServerBusy",
                "CodeNeedLogin",
                "CodeInvalidToken"
            ]
        },
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/controller.ResCode"
                        }
                    ]
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "提示信息"
                }
            }
        },
        "models.ApiPostDetail": {
            "type": "object",
            "properties": {
                "author_name": {
                    "description": "作者用户名",
                    "type": "string"
                },
                "community": {
                    "description": "所属社区信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CommunityDetail"
                        }
                    ]
                },
                "post": {
                    "description": "帖子详情",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Post"
                        }
                    ]
                },
                "vote_num": {
                    "description": "赞成票数",
                    "type": "integer"
                }
            }
        },
        "models.Community": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "社区ID",
                    "type": "integer"
                },
                "name": {
                    "description": "社区名称",
                    "type": "string"
                }
            }
        },
        "models.CommunityDetail": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "社区ID",
                    "type": "integer"
                },
                "introduction": {
                    "description": "社区简介",
                    "type": "string"
                },
                "name": {
                    "description": "社区名称",
                    "type": "string"
                }
            }
        },
        "models.LoginData": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT Token",
                    "type": "string",
                    "example": "eyJhbGciOi..."
                },
                "user_id": {
                    "description": "用户ID（字符串格式，避免JS精度丢失）",
                    "type": "string",
                    "example": "1234567890"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
        "models.LoginParam": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2
                }
            }
        },
        "models.ParamVoteData": {
            "type": "object",
            "required": [
                "post_id"
            ],
            "properties": {
                "direction": {
                    "description": "投票方向：1=赞成，-1=反对，0=取消",
                    "type": "string",
                    "enum": [
                        1,
                        -1,
                        0
                    ],
                    "example": "0"
                },
                "post_id": {
                    "description": "帖子ID",
                    "type": "string"
                }
            }
        },
        "models.Post": {
            "type": "object",
            "required": [
                "community_id",
                "content",
                "title"
            ],
            "properties": {
                "author_id": {
                    "description": "作者用户ID",
                    "type": "integer"
                },
                "community_id": {
                    "description": "所属社区ID",
                    "type": "integer"
                },
                "content": {
                    "description": "帖子内容",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "帖子ID（雪花算法生成）",
                    "type": "integer"
                },
                "status": {
                    "description": "帖子状态",
                    "type": "integer"
                },
                "title": {
                    "description": "帖子标题",
                    "type": "string"
                }
            }
        },
        "models.SignUpData": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "提示信息",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.SignUpParam": {
            "type": "object",
            "required": [
                "password",
                "re_password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码，6-20个字符",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "re_password": {
                    "description": "确认密码，需与密码一致",
                    "type": "string"
                },
                "username": {
                    "description": "用户名，2-20个字符",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2
                }
            }
        },
        "models.VoteData": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}